Add redirect to home page when user is logged in

+++ base/views.py
@@ -11,6 +11,10 @@ from django.contrib.auth import authenticate, login, logout
 
 
 def loginView(request):
+
+    if request.user.is_authenticated:
+        return redirect('home')
+
     if request.method == 'POST':
         username = request.POST.get('username')
         password = request.POST.get('password')
@@ -62,6 +66,9 @@ def room(request, pk):
 def createRoom(request):
     form = RoomForm()
 
+    if request.user != room.host:
+        return HttpResponse('Your are not allowed here!')
+
     if request.method == 'POST':
         form = RoomForm(request.POST)
         if form.is_valid():
@@ -74,9 +81,13 @@ def createRoom(request):
 
     return render(request, 'base/room_form.html', context )
 
+@login_required(login_url='/login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse('Your are not allowed here!')
 
     if request.method == 'POST':
         form =  RoomForm(request.POST, instance=room)
@@ -89,8 +100,13 @@ def updateRoom(request, pk):
     }
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='/login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse('Your are not allowed here!')
+
     if request.method == 'POST':
         room.delete()
         return redirect('home')
